+++ src/Colors/Cmyk/Color.php
@@ -161,9 +161,9 @@ class Color extends AbstractColor
     /**
      * {@inheritdoc}
      *
-     * @see ColorInterface::isGreyscale()
+     * @see ColorInterface::isGrayscale()
      */
-    public function isGreyscale(): bool
+    public function isGrayscale(): bool
     {
         return 0 === array_sum([
             $this->cyan()->value(),