+++ src/Colors/Hsl/Color.php
@@ -158,9 +158,9 @@ class Color extends AbstractColor
     /**
      * {@inheritdoc}
      *
-     * @see ColorInterface::isGreyscale()
+     * @see ColorInterface::isGrayscale()
      */
-    public function isGreyscale(): bool
+    public function isGrayscale(): bool
     {
         return $this->saturation()->value() == 0;
     }